requirements.txt
flask
pytest
requests
pyyaml
responses